+++ geowatch/tasks/fusion/predict.py
@@ -560,9 +560,10 @@ def _predict_critical_loop(config, fit_config, model, datamodule, result_dataset
     print('Predict on device = {!r}'.format(device))
     downweight_edges = config.downweight_edges
 
-    UNPACKAGE_METHOD_HACK = 0
+    UNPACKAGE_METHOD_HACK = 1
     if UNPACKAGE_METHOD_HACK:
-        # unpackage model hack
+        # unpackage model hack, this only works if we know the model that will
+        # be loaded a-priori
         from geowatch.tasks.fusion import methods
         unpackged_method = methods.MultimodalTransformer(**model.hparams)
         unpackged_method.load_state_dict(model.state_dict())